Make point motor loading a bit easier to understand

+++ src/arduino/objects/motors/PointsMotor_Pulse.h
@@ -21,9 +21,9 @@ class PointsMotor_Pulse: public PointsMotor {
     virtual int objectInit(const uint8_t configData[]);
     virtual IPOCS::PointsStatusPacket::E_RQ_POINTS_STATE getState();
   private:
-    int throwLeftOutput;
-    int throwRightOutput;
-    int posInput;
+    uint8_t throwLeftOutput;
+    uint8_t throwRightOutput;
+    uint8_t posInput;
 
     unsigned long lastOrderMillis;
     IPOCS::ThrowPointsPacket::E_RQ_POINTS_COMMAND lastOrderState;
